serialization: allow longer messages
Shift the encoded length value in the header left by 1, so the maximum
allowed message length doubles from 255 to 510.

+++ src/serialize/create.rs
@@ -20,10 +20,10 @@ pub struct Msg {
 }
 
 impl AsRawMsg for Msg {
-    fn get_hdr(&self) -> (u8, u8, u32) {
+    fn get_hdr(&self) -> (u8, u32, u32) {
         (
             CREATE,
-            HDR_LENGTH + 6 * 4 + self.cong_alg.len() as u8,
+            HDR_LENGTH + 6 * 4 + self.cong_alg.len() as u32,
             self.sid,
         )
     }
