+++ src/serialize/install_fold.rs
@@ -15,10 +15,10 @@ pub struct Msg {
 }
 
 impl AsRawMsg for Msg {
-    fn get_hdr(&self) -> (u8, u8, u32) {
+    fn get_hdr(&self) -> (u8, u32, u32) {
         (
             INSTALL_FOLD,
-            HDR_LENGTH + 4 + (self.instrs.0.len() * 10) as u8,
+            HDR_LENGTH + 4 + (self.instrs.0.len() as u32 * 10),
             self.sid,
         )
     }
